+++ swagger-request-validator-core/src/main/java/com/atlassian/oai/validator/interaction/RequestValidator.java
@@ -59,17 +59,14 @@ public class RequestValidator {
     /**
      * Validate the request against the given API operation
      *
-     * @param requestPath The normalised path the request is on
      * @param request The request to validate
      * @param apiOperation The operation to validate the request against
      *
      * @return A validation report containing validation errors
      */
     @Nonnull
-    public ValidationReport validateRequest(@Nonnull final NormalisedPath requestPath,
-                                            @Nonnull final Request request,
+    public ValidationReport validateRequest(@Nonnull final Request request,
                                             @Nonnull final ApiOperation apiOperation) {
-        requireNonNull(requestPath, "A request path is required");
         requireNonNull(request, "A request is required");
         requireNonNull(apiOperation, "An API operation is required");
 
@@ -77,7 +74,7 @@ public class RequestValidator {
                 .merge(validateContentType(request, apiOperation))
                 .merge(validateAccepts(request, apiOperation))
                 .merge(validateHeaders(request, apiOperation))
-                .merge(validatePathParameters(requestPath, apiOperation))
+                .merge(validatePathParameters(apiOperation))
                 .merge(validateRequestBody(request.getBody(), apiOperation))
                 .merge(validateQueryParameters(request, apiOperation));
     }
@@ -234,13 +231,19 @@ public class RequestValidator {
                                           @Nonnull final ApiOperation apiOperation) {
 
         final Multimap<String, String> formData = parseFormData(requestBody.get());
-        final List<ValidationReport> reports = new ArrayList<>();
-        for (Parameter parameter : apiOperation.getOperation().getParameters()) {
-            Collection<String> parameterValues = formData.get(parameter.getName());
-            parameterValues = parameterValues.isEmpty() ? Collections.singletonList(null) : parameterValues;
-            parameterValues.forEach(value -> reports.add(parameterValidators.validate(value, parameter)));
-        }
-        return reports.stream().reduce(ValidationReport.empty(), ValidationReport::merge);
+        return apiOperation.getOperation().getParameters().stream()
+                .flatMap(parameter ->
+                        prepareFormDataForParameter(formData, parameter).stream()
+                                .map(value -> parameterValidators.validate(value, parameter))
+                )
+                .reduce(ValidationReport.empty(), ValidationReport::merge);
+    }
+
+    @Nonnull
+    private Collection<String> prepareFormDataForParameter(@Nonnull final Multimap<String, String> formData,
+                                                           @Nonnull final Parameter parameter) {
+        final Collection<String> parameterValues = formData.get(parameter.getName());
+        return parameterValues.isEmpty() ? Collections.singletonList(null) : parameterValues;
     }
 
     @Nonnull
@@ -274,11 +277,11 @@ public class RequestValidator {
     }
 
     @Nonnull
-    private ValidationReport validatePathParameters(@Nonnull final NormalisedPath requestPath,
-                                                    @Nonnull final ApiOperation apiOperation) {
+    private ValidationReport validatePathParameters(@Nonnull final ApiOperation apiOperation) {
 
         ValidationReport validationReport = ValidationReport.empty();
-        for (int i = 0; i < apiOperation.getPathString().parts().size(); i++) {
+        final NormalisedPath requestPath = apiOperation.getRequestPath();
+        for (int i = 0; i < apiOperation.getPathString().numberOfParts(); i++) {
             if (!apiOperation.getPathString().isParam(i)) {
                 continue;
             }