+++ swagger-request-validator-core/src/main/java/com/atlassian/oai/validator/parameter/ArrayParameterValidator.java
@@ -1,7 +1,6 @@
 package com.atlassian.oai.validator.parameter;
 
 import com.atlassian.oai.validator.report.MessageResolver;
-import com.atlassian.oai.validator.report.MutableValidationReport;
 import com.atlassian.oai.validator.report.ValidationReport;
 import com.atlassian.oai.validator.schema.SchemaValidator;
 import io.swagger.models.parameters.Parameter;
@@ -14,6 +13,7 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.Set;
+import java.util.stream.Stream;
 
 import static java.util.Objects.requireNonNull;
 
@@ -69,98 +69,102 @@ public class ArrayParameterValidator extends BaseParameterValidator {
     @Override
     @Nonnull
     public ValidationReport validate(@Nullable final String value, @Nullable final Parameter p) {
-        final MutableValidationReport report = new MutableValidationReport();
-
         if (!supports(p)) {
-            return report;
+            return ValidationReport.EMPTY_REPORT;
         }
 
         final SerializableParameter parameter = (SerializableParameter) p;
 
         if (parameter.getRequired() && (value == null || value.trim().isEmpty())) {
-            return report.add(messages.get("validation.request.parameter.missing", parameter.getName()));
+            return ValidationReport.singleton(messages.get("validation.request.parameter.missing", parameter.getName()));
         }
 
         if (value == null || value.trim().isEmpty()) {
-            return report;
+            return ValidationReport.EMPTY_REPORT;
         }
 
-        doValidate(value, parameter, report);
-        return report;
+        return doValidate(value, parameter);
     }
 
     public ValidationReport validate(@Nullable final Collection<String> values, @Nullable final Parameter p) {
-        final MutableValidationReport report = new MutableValidationReport();
         if (p == null) {
-            return report;
+            return ValidationReport.EMPTY_REPORT;
         }
 
         final SerializableParameter parameter = (SerializableParameter) p;
         if (parameter.getRequired() && (values == null || values.isEmpty())) {
-            return report.add(messages.get("validation.request.parameter.missing", parameter.getName()));
+            return ValidationReport.singleton(messages.get("validation.request.parameter.missing", parameter.getName()));
         }
 
         if (values == null) {
-            return report;
+            return ValidationReport.EMPTY_REPORT;
         }
 
         if (!parameter.getCollectionFormat().equalsIgnoreCase(CollectionFormat.MULTI.name())) {
-            return report.add(messages.get("validation.request.parameter.collection.invalidFormat",
+            return ValidationReport.singleton(messages.get("validation.request.parameter.collection.invalidFormat",
                     p.getName(), parameter.getCollectionFormat(), "multi")
             );
         }
 
-        doValidate(values, parameter, report);
-        return report;
+        return doValidate(values, parameter);
     }
 
     @Override
-    protected void doValidate(@Nonnull final String value,
-                              @Nonnull final SerializableParameter parameter,
-                              @Nonnull final MutableValidationReport validationReport) {
+    protected ValidationReport doValidate(@Nonnull final String value,
+                                          @Nonnull final SerializableParameter parameter) {
 
-        doValidate(CollectionFormat.from(parameter).split(value),
-                parameter,
-                validationReport);
-
+        return doValidate(CollectionFormat.from(parameter).split(value), parameter);
     }
 
-    private void doValidate(@Nonnull final Collection<String> values,
-                            @Nonnull final SerializableParameter parameter,
-                            @Nonnull final MutableValidationReport validationReport) {
+    private ValidationReport doValidate(@Nonnull final Collection<String> values,
+                                        @Nonnull final SerializableParameter parameter) {
 
-        if (parameter.getMaxItems() != null && values.size() > parameter.getMaxItems()) {
-            validationReport.add(messages.get("validation.request.parameter.collection.tooManyItems",
-                    parameter.getName(), parameter.getMaxItems(), values.size())
-            );
-        }
-
-        if (parameter.getMinItems() != null && values.size() < parameter.getMinItems()) {
-            validationReport.add(messages.get("validation.request.parameter.collection.tooFewItems",
-                    parameter.getName(), parameter.getMinItems(), values.size())
-            );
-        }
-
-        if (Boolean.TRUE.equals(parameter.isUniqueItems()) &&
-                values.stream().distinct().count() != values.size()) {
-            validationReport.add(messages.get("validation.request.parameter.collection.duplicateItems",
-                    parameter.getName())
-            );
-        }
+        final ValidationReport report = Stream.of(
+                validateMaxItems(values, parameter),
+                validateMinItems(values, parameter),
+                validateUniqueItems(values, parameter)
+        ).reduce(ValidationReport.empty(), ValidationReport::merge);
 
         if (parameter.getEnum() != null && !parameter.getEnum().isEmpty()) {
             final Set<String> enumValues = new HashSet<>(parameter.getEnum());
-            values.stream()
+            return values.stream()
                     .filter(v -> !enumValues.contains(v))
-                    .forEach(v -> {
-                        validationReport.add(messages.get("validation.request.parameter.enum.invalid",
-                                v, parameter.getName(), parameter.getEnum())
-                        );
-                    });
-            return;
+                    .map(v -> ValidationReport.singleton(messages.get("validation.request.parameter.enum.invalid",
+                            v, parameter.getName(), parameter.getEnum())
+                    ))
+                    .reduce(report, ValidationReport::merge);
         }
 
-        values.forEach(v ->
-                validationReport.addAll(schemaValidator.validate(v, parameter.getItems())));
+        return values.stream()
+                .map(v -> schemaValidator.validate(v, parameter.getItems()))
+                .reduce(report, ValidationReport::merge);
+    }
+
+    private ValidationReport validateUniqueItems(final @Nonnull Collection<String> values, final @Nonnull SerializableParameter parameter) {
+        if (Boolean.TRUE.equals(parameter.isUniqueItems()) &&
+            values.stream().distinct().count() != values.size()) {
+            return ValidationReport.singleton(messages.get("validation.request.parameter.collection.duplicateItems",
+                parameter.getName())
+            );
+        }
+        return ValidationReport.empty();
+    }
+
+    private ValidationReport validateMinItems(final @Nonnull Collection<String> values, final @Nonnull SerializableParameter parameter) {
+        if (parameter.getMinItems() != null && values.size() < parameter.getMinItems()) {
+            return ValidationReport.singleton(messages.get("validation.request.parameter.collection.tooFewItems",
+                parameter.getName(), parameter.getMinItems(), values.size())
+            );
+        }
+        return ValidationReport.empty();
+    }
+
+    private ValidationReport validateMaxItems(final @Nonnull Collection<String> values, final @Nonnull SerializableParameter parameter) {
+        if (parameter.getMaxItems() != null && values.size() > parameter.getMaxItems()) {
+            return ValidationReport.singleton(messages.get("validation.request.parameter.collection.tooManyItems",
+                parameter.getName(), parameter.getMaxItems(), values.size())
+            );
+        }
+        return ValidationReport.empty();
     }
 }