+++ swagger-request-validator-core/src/main/java/com/atlassian/oai/validator/parameter/format/FormatValidator.java
@@ -1,10 +1,10 @@
 package com.atlassian.oai.validator.parameter.format;
 
-import com.atlassian.oai.validator.report.MutableValidationReport;
+import com.atlassian.oai.validator.report.ValidationReport;
 
 public interface FormatValidator<T> {
 
     boolean supports(String format);
 
-    void validate(MutableValidationReport report, T value);
+    ValidationReport validate(T value);
 }